requirements.txt
# fastapi
# uvicorn
# sqlalchemy
# psycopg2-binary # PostgreSQL 数据库驱动
# pydantic
# passlib[bcrypt]
# python-jose[jwt]
# python-dotenv
# email-validator 
# python-multipart

fastapi==0.111.0
uvicorn[standard]==0.29.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic==2.7.4